GUACAMOLE-573: Move terminal text selection code into own file.
